requirements.txt
numpy
gin-config
pygame
h5py
pyzmq